requirements.txt
Pillow==8.1.0
pyaes==1.6.1
pyasn1==0.4.8
pytz==2020.5
rsa==4.7
Telethon==1.19.0
